Create new R Markdown package

Populates the directory specified via path with all the necessary files for a new R Markdown package.

The DESCRIPTION file is created using usethis::use_description() and all fields except Package, URL and BugReports are sourced from the usethis.description R option if defined.

Usage

create_pkg(
  name,
  id_netlify = NULL,
  path = ".",
  incl_roxygen2_meta = TRUE,
  incl_reexports = FALSE,
  incl_sysdata_rmd = FALSE,
  incl_data_rmd = FALSE,
  incl_asciicasts_rmd = FALSE,
  incl_pkgdown_config = TRUE,
  incl_ripgrep_config = TRUE,
  incl_ack_config = FALSE,
  incl_air_config = FALSE
)

Arguments

name

Package name.

id_netlify

Netlify site identifier.

path

Path to the new package directory.

incl_roxygen2_meta

Whether or not to create a man/roxygen/meta.R file that i.a. stores the metadata for roxygen2's @family tags.

incl_reexports

Whether or not to create an R/reexports.R file prefilled with an opinionated set of magrittr and rland operator exports.

incl_sysdata_rmd

Whether or not to create an Rmd/sysdata.nopurl.Rmd stub file.

incl_data_rmd

Whether or not to create an Rmd/data.nopurl.Rmd stub file.

incl_asciicasts_rmd

Whether or not to create an Rmd/asciicasts.nopurl.Rmd stub file.

incl_pkgdown_config

Whether or not to create a minimal pkgdown/_pkgdown.yml configuration file.

incl_ripgrep_config

Whether or not to create a ripgrep ignore file .rgignore.

incl_ack_config

Whether or not to create an .ackrc configuration file.

incl_air_config

Whether or not to create an air.toml configuration file. Note that air is not of much use yet for R Markdown packages.

Value

path, invisibly.