Generate pkgdown reference index

Generates the pkgdown reference index based on the heading hierarchy structure of a package's main .Rmd file.

gen_pkgdown_ref(rmd)

Arguments

rmd
The (R) Markdown file content as a character scalar.

Value

A list.

Details

Basically, all of the pkgdown reference index except desc keys are derived from rmd's Markdown headings and their hierarchy. The desc keys, however, are assembled from the text paragraphs below some specially named headings (see below).

To be able to unambiguously map an .Rmd file's heading hierarchy to the pkgdown reference index, some rules are necessary. Attention must be paid to the fact that, while (R) Markdown supports up to six heading levels (corresponding to HTML's <h1>–<h6> tags), the pkgdown reference index only supports up to two (title and subtitle).

The following rules define how the reference index is generated:

Headings below a heading named INTERNAL, TEMP or TEMPORARY (case-insensitive, but without any inline formatting) are simply ignored when generating the reference index.

Every heading that is a) inline-formatted as verbatim and b) doesn't contain any whitespace characters is considered to be the name of a help topic (usually the name of a function or dataset) to be included in the reference index. This maps to the contents key of the reference index' YAML.

Non-help-topic headings above help topic headings are used as reference index (sub)titles as far as hierarchical nesting allows. More precisely, non-help-topic headings of the highest two levels above the help topic heading are used as title and subtitle, the rest of the headings above the help topic heading is ignored. This maps to the title and subtitle keys of the reference index' YAML.

Text paragraph(s) below headings named DESCRIPTION (case-insensitive, but without any inline formatting), that in turn are hierarchically below a title or subtitle heading, will be used as descriptions for the (sub)titles. This maps to the desc key of the reference index' YAML.

When fed to gen_pkgdown_ref(), the following simplified example .Rmd...

---
editor_options:
  markdown:
    wrap: 80
---

# `main_fn`

The fn below is under no title or subtitle in the pkgdown index.

```{r}
main_fn <- function(...) {...}
```

# First group of fns

## DESCRIPTION

The description text for *First group of fns*.

## `group_wide_fn`

The fn below is only under a title, but no subtitle in the pkgdown index.

```{r}
group_wide_fn <- function(...) {...}
```

## Some subgroup

The fns below are both under a title and subtitle in the pkgdown index.

### `subgroup_fn_1`

```{r}
subgroup_fn_1 <- function(...) {...}
```

### `subgroup_fn_2`

```{r}
subgroup_fn_2 <- function(...) {...}
```

### Sub-subgroup

The above *Sub-subgroup* heading is ignored for the pkgdown index because the
latter only supports up to two heading levels.

The fn below thus moves one hierarchy level up in the pkgdown index.

#### `sub_subgroup_fn`

```{r}
sub_subgroup_fn <- function(...) {...}
```

## Another subgroup

### DESCRIPTION

The description text for *Another subgroup*.

### `another_subgroup_fn`

```{r}
another_subgroup_fn <- function(...) {...}
```

# Second group of fns

## Yet another subgroup

### `yet_another_subgroup_fn`

```{r}
yet_another_subgroup_fn <- function(...) {...}
```

...yields this pkgdown index YAML:

reference:
- contents: main_fn
- title: First group of fns
  desc: The description text for *First group of fns*.
- contents: group_wide_fn
- subtitle: Some subgroup
- contents:
  - subgroup_fn_1
  - subgroup_fn_2
  - sub_subgroup_fn
- subtitle: Another subgroup
  desc: The description text for *Another subgroup*.
- contents: another_subgroup_fn
- title: Second group of fns
- subtitle: Yet another subgroup
- contents: yet_another_subgroup_fn

Examples

pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
  pkgpurl::gen_pkgdown_ref() |>
  yaml::as.yaml() |>
  cat()

#> reference:
#> - title: Statistical computing / numbers
#> - contents:
#>   - safe_seq_len
#>   - safe_max
#>   - safe_min
#>   - round_to
#>   - stat_mode
#> - title: Data frames / Tibbles
#> - contents:
#>   - is_equal_df
#>   - reduce_df_list
#> - title: Lists
#> - contents:
#>   - as_flat_list
#>   - list_drop
#>   - list_keep
#> - title: Strings
#> - contents:
#>   - as_comment_string
#>   - as_chr
#>   - as_line_feed_chr
#>   - as_string
#>   - sentenceify
#>   - escape_lf
#>   - fuse_regex
#>   - dsv_colnames
#>   - prose_ls
#>   - prettify_nr
#>   - cat_lines
#>   - wrap_chr
#> - title: Dots
#>   desc: |-
#>     Extending the [ellipsis](https://ellipsis.r-lib.org/) package, making the use of [R's `...` argument placeholder](https://rdrr.io/r/base/dots.html) yet another
#>     bit safer.
#> - contents: check_dots_named
#> - title: R packages
#> - contents:
#>   - ls_pkg
#>   - assert_pkg
#>   - is_pkg_installed
#>   - is_pkg_cran
#>   - is_pkg_dir
#>   - is_pkgdown_dir
#> - subtitle: Package DESCRIPTION
#>   desc: Extending the [desc](https://github.com/r-lib/desc#readme) package.
#> - contents:
#>   - desc_list
#>   - desc_value
#>   - desc_url_git
#> - subtitle: Package documentation
#>   desc: Extending the [roxygen2](https://roxygen2.r-lib.org/) package.
#> - contents:
#>   - roxy_blocks
#>   - roxy_obj
#>   - roxy_tag_value
#> - title: (Pandoc) Markdown
#> - contents:
#>   - as_md_list
#>   - pipe_table
#>   - strip_md
#>   - strip_md_footnotes
#> - subtitle: CommonMark parsing
#>   desc: Extending the [commonmark](https://github.com/r-lib/commonmark#readme) package.
#> - contents:
#>   - md_to_xml
#>   - md_xml_subnode_ix
#>   - xml_to_md
#> - title: R Markdown / knitr
#> - contents:
#>   - build_readme
#>   - knitr_table_format
#>   - strip_yaml_header
#> - subtitle: Output formats
#>   desc: |-
#>     [Custom R Markdown output formats](https://bookdown.org/yihui/rmarkdown/new-formats.html) which can be used in addition to the [default output
#>     formats](https://bookdown.org/yihui/rmarkdown/output-formats.html).
#> - contents: gitlab_document
#> - title: GitHub
#>   desc: Extending the [gh](https://gh.r-lib.org/) package, leveraging [GitHub's GraphQL
#>     API](https://docs.github.com/en/graphql).
#> - contents:
#>   - gh_text_file
#>   - gh_text_files
#>   - gh_dir_ls
#> - title: HTTP
#>   desc: Extending the [httr](https://httr.r-lib.org/) package that wraps the [curl](https://jeroen.cran.dev/curl/)
#>     package.
#> - contents:
#>   - assert_mime_type
#>   - is_http_success
#> - title: System interaction
#> - contents:
#>   - check_cli
#>   - path_script
#> - title: Miscellaneous
#> - contents:
#>   - assert_class_any
#>   - capture_print
#>   - cols_regex
#>   - cli_process_expr
#>   - order_by
#>   - prose_ls_fn_param