Lint R Markdown package

This is a convenience wrapper around lintr::lint_dir() which is tailored to a typical R Markdown package. To use this function, the lintr package is required.

lint_rmd(path = ".", exclude_vignettes = TRUE)

Arguments

path
The path to the root of the package directory.

exclude_vignettes
Whether to exclude all .Rmd files under vignettes/. A logical scalar.

Details

To avoid unnecessary noise, all the the generated R/*.gen.R files as well as R Markdown vignettes under vignettes/*.Rmd are excluded from linting.

This function is also registered as an RStudio add-in allowing RStudio users to assign a custom shortcut to it.