Lint R Markdown package

This is a convenience wrapper around lintr::lint_dir() which is tailored to a typical R Markdown package. To use this function, the lintr package is required.

Usage

lint_rmd(path = ".", excl_vignettes = TRUE)

Arguments

path

Path to the root of the package directory.

excl_vignettes

Whether to exclude all .Rmd files under vignettes/. A logical scalar.

Value

A list of lint objects.

Details

To avoid unnecessary noise, all the the generated R/*.gen.R files as well as R Markdown vignettes under vignettes/*.Rmd are excluded from linting.

This function is also registered as an RStudio add-in, allowing RStudio users to assign a custom shortcut to it or to invoke it from the command palette.

See also